Add Coinbase balance and outstanding order balances to orders page
This change adds a simple table to the orders page the displays 1) your current BTC balance on Coinbase, 2) the total of sell orders you have outstanding, 3) the difference between 1 and 2, 4) the total of buy orders you have outstanding.

The Coinbase balance incurs an API request to Coinbase, but it is unlikely to change frequently, so it is a good candidate for caching.

Test Plan:
Login and visit /homepage/orders
You should see a table with the 4 pieces of information described above.