Add global stats to the homepage
I resolved the previous issue with some advice from Tom, which means that all of the global stats values are not properly being computed. I added some code to route these through to the homepage and display them, but that will only work once the cron job has fired.  This shouldn't be too big of a deal on production since the cron job constantly fires.

If memcache is available, caching the global stats there would be ideal.

Test Plan:
Visit /api/cron, to make sure the values are populated
Visit the logged out homepage to see stats displayed there